[Manifest] Close the gaps between the two enforcers, and fold tests/perf into the op tests
The roofline codegen accepted an optional input's `.shape` and `[...]` that
the validator rejects, so the two enforcers disagreed on what a `vars` entry
may read; the codegen now rejects both.

Dtype parity stopped probing an optional input entirely once it left the combo
rows, dropping the coverage those tensors had while they were required. It now
probes each listed combo again with every optional passed at a declared dtype.

`optional` is rejected on a param, which expresses optionality with `default`.
`same_as(X)` on an optional input no longer has a self-reference exception: an
absent input has no dtype to resolve to, whoever asks. The generated
`_validate_dtypes` makes optional inputs keyword-only, so declaring one before
a required input no longer emits a signature Python cannot parse.

InstanceNorm's roofline counted running-stats bytes on every call that passed
them. The batch-stats path never reads them and no path writes them back, so
it counts the two reads of the running-stats path alone.

tests/perf held roofline contracts for five op families in files of their own.
Each moves into the test file for the op it describes, and the profile test
moves beside the other module tests. Node count is unchanged.

Comments and error strings say the fact instead of a rule number, and the
manifest no longer repeats a global rule in a family's comment.
