+++ src/tileops/manifest/attention_indexing.yaml
@@ -33,8 +33,6 @@ FP8LightningIndexerOp:
   workloads:
   - {batch: 1, seq_len: 8192, heads: 32, index_dim: 64, seq_len_kv: 8192, kv_group: 1, clean_logits: true, dtypes: [bfloat16], label: "lightning-indexer-bf16-s8k-h32-d64"}
   - {batch: 1, seq_len: 8192, heads: 32, index_dim: 64, seq_len_kv: 8192, kv_group: 1, clean_logits: true, dtypes: [float8_e4m3fn], label: "lightning-indexer-fp8-s8k-h32-d64"}
-  # A row passes index_k_scale by carrying index_k_scale_shape, whatever else
-  # it writes; the rows above give shapes as scalar dims.
   - {batch: 1, seq_len: 8192, heads: 32, index_dim: 64, seq_len_kv: 8192, kv_group: 1, clean_logits: true, dtypes: [float8_e4m3fn], index_k_scale_shape: [1, 8192, 1], label: "lightning-indexer-fp8-s8k-h32-d64-kscale"}
 
   roofline:
