+++ src/tileops/manifest/normalization.yaml
@@ -511,13 +511,14 @@ InstanceNormFwdOp:
       C: "x.shape[1]"
       spatial_size: "product(x.shape[2:])"
       affine: "weight is not None"
-      tracks_stats: "running_mean is not None"
+      # The batch-stats path ignores them, and no path writes them back.
+      reads_running_stats: "running_mean is not None and not use_input_stats"
     # Per element: subtract mean + square for var + normalize = 3; affine adds
     # scale + bias, matching GroupNorm (InstanceNorm = GroupNorm with G=C).
     flops: "(5 if affine else 3) * N * C * spatial_size"
     # Read x + write output; affine reads weight and bias (C each); passed
     # running stats are read and written back, 2 * C float32 each way.
-    bytes: "(2 * N * C * spatial_size + (2 * C if affine else 0)) * elem_bytes + (4 * C * 4 if tracks_stats else 0)"
+    bytes: "(2 * N * C * spatial_size + (2 * C if affine else 0)) * elem_bytes + (2 * C * 4 if reads_running_stats else 0)"
 
   source:
     kernel: tileops/kernels/norm/group_norm.py
