+++ tests/ops/test_gated_deltanet_prefill.py
@@ -3,6 +3,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import GatedDeltaNetPrefillFwdOp
+from tileops.perf.formulas import gated_deltanet_prefill_fwd_roofline
 from workloads.linear_attention import (
     GatedDeltaNetPrefillFwdWorkload,
     compute_w_u_torch,
@@ -373,3 +374,30 @@ def test_gated_deltanet_prefill_dense_transition_reference() -> None:
 
 if __name__ == "__main__":
     pytest.main([__file__, "-vvs"])
+
+
+# ----------------------------------------------------------------------
+# Cross-layout contract for the Gated DeltaNet prefill roofline.
+# ----------------------------------------------------------------------
+
+
+
+@pytest.mark.smoke
+def test_gated_deltanet_prefill_roofline_layout_equivalence() -> None:
+    """bthd and bhtd bindings of the same problem yield identical costs."""
+    bthd = gated_deltanet_prefill_fwd_roofline(
+        q_shape=[1, 512, 16, 128],
+        v_shape=[1, 512, 16, 128],
+        chunk_size=64,
+        layout="bthd",
+        dtype="float16",
+    )
+    bhtd = gated_deltanet_prefill_fwd_roofline(
+        q_shape=[1, 16, 512, 128],
+        v_shape=[1, 16, 512, 128],
+        chunk_size=64,
+        layout="bhtd",
+        dtype="float16",
+    )
+    assert bthd == bhtd
+    assert bthd[0] > 0 and bthd[1] > 0
